Roughly fix the management of Replication Lookup Ctl table (rctl).
The original implementation of PGR only allows ~1000 coid/xid stored
in (rctl). There is only store functions, but not erase functions (how
stupied is that). After the empty slots run out, continue to call
store function will evevntually overwrite the memory of next share
memory segement -- which in current PGR is Buffer Descriptors. That's
why there will be blood on get_database_list!

This rough modification uses following strategy:

1. a new field valid is added to PeerTxnEntry, to indicate whether
this slot is still valid --- after commitment or abortion, I think it
is useless and make it in-valid.

2. allocation new entry of rctl is now managed with function
alloc_new_entry, which will (i) allocate a new entry in the head (if
there is empty space); (ii) or scan the rctl list, remove the invalid
entries, and then allocate. I hope this strategy can somehow amoritize
the cost. But there is definitely room to further improve this part.