Fix a lot in the management of Replication Lookup Ctl tabl (rctl).
Last time rough fix is not right: the xid and coid of committed
transaction can not be simply thrown out of the memmory, it will cause
some transaction still rely on it do not know what to do.

Therefore, we revise again as follows

1. A system table pg_replication has been setup as

CREATE TABLE pg_replication (
       origin_local_node_id int32,
       origin_local_xid int32,
       local_xid int32,
       local_coid int32
)

And there are hash indexes

ReplicationOriginIndex on (origin_local_node_id, origin_local_xid)
ReplicationXidIndex on (local_xid)
ReplicationCoidIndex on (local_coid)

2. The list in shmem is now managed with a clock replacement
policy. And we have added three hash lookup table

HTAB           *PteOriginHash;
HTAB           *PteLocalXidHash;
HTAB           *PteLocalCoidHash;

3. Committed or aborted transaction will be marked valid from shmem
list. And when shmem list is short of slots, the in-valid slot will be
reused by following clock replacement policy. Before resuing the slot,

if the slot is for a committed transaction, store it to pg_replication
otherwise if the slot is for an aborted transaction, no store.

4. Now query functions have all added with the ability to query
pg_replication in case not find in shmem.